refactor(consumers): rewrite 6 source consumers to import from core/ports
Rewrites all non-stt/tts source files that imported from lyra.stt or
lyra.tts to use lyra.core.ports.{stt,tts} instead. Covers middleware,
agent base, simple_agent, simple_agent_prompts (lazy import), and both
bootstrap factory files. Also drops 3 stale ignore_imports entries from
.importlinter that referenced the now-rewritten consumer imports.
